Refactor: move reducers closer to components
