+++ pages/experiment/[experimentid].tsx
@@ -1,9 +1,9 @@
 import { useRouter } from 'next/router'
-import { ExperimentProvider } from '../../src/context/experiment-context'
+import { ExperimentProvider } from '../../src/context/experiment'
 import TabbedExperiment from '../../src/components/experiment/tabbed-experiment'
 import Experiment from '../../src/components/experiment/experiment'
-import DebugExperiment from '../../src/components/debug-experiment'
-import { useGlobal } from '../../src/context/global-context'
+import DebugExperiment from '../../src/components/experiment/debug-experiment'
+import { useGlobal } from '../../src/context/global'
 import LoadingExperiment from '../../src/components/experiment/loading-experiment'
 import JsonEditor from '../../src/components/json-editor/json-editor'
