+++ src/components/experiment/dataEntryTab.tsx
@@ -1,5 +1,5 @@
 import { Grid } from '@mui/material'
-import { useExperiment } from '../../context/experiment-context'
+import { useExperiment } from '../../context/experiment'
 import { DataPointType } from '../../types/common'
 import DataPoints from '../data-points/data-points'
 import { ExperimentationGuide } from '../result-data/experimentation-guide'
