+++ src/components/experiment/debug-experiment.tsx
@@ -1,5 +1,5 @@
 import { Card, CardContent } from '@mui/material'
-import { useExperiment } from '../context/experiment-context'
+import { useExperiment } from '../../context/experiment'
 
 export default function DebugExperiment() {
   const { state } = useExperiment()
