+++ src/components/experiment/tabbed-experiment.tsx
@@ -16,17 +16,17 @@ import {
   useExperiment,
   runExperiment,
   useSelector,
-} from '../../context/experiment-context'
+} from '../../context/experiment'
 import React, { useState } from 'react'
 import LoadingExperiment from './loading-experiment'
 import LoadingButton from '../loading-button/loading-button'
 import { Plots } from '../plots/plots'
 import { saveObjectToLocalFile } from '../../utility/save-to-local-file'
-import { useGlobal } from '../../context/global-context'
+import { useGlobal } from '../../context/global'
 import { ConfigurationTab } from './configurationTab'
 import { DataEntryTab } from './dataEntryTab'
-import { State } from '../../reducers/global-reducer'
-import { selectIsInitializing } from '../../reducers/experiment-selectors'
+import { State } from '../../context/global'
+import { selectIsInitializing } from '../../context/experiment'
 
 type SnackbarMessage = {
   message: string
