+++ src/components/json-editor/json-editor.tsx
@@ -8,11 +8,11 @@ import {
   Typography,
 } from '@mui/material'
 import { ChangeEvent, useEffect, useState } from 'react'
-import { useExperiment } from '../../context/experiment-context'
+import { useExperiment } from '../../context/experiment'
 import useStyles from './json-editor.style'
 import { ExperimentType } from '../../types/common'
 import CloseIcon from '@mui/icons-material/Close'
-import { useGlobal } from '../../context/global-context'
+import { useGlobal } from '../../context/global'
 import { errorMessage } from '../../utility'
 
 type DisplayedResults = {
