+++ src/components/plots/plots.tsx
@@ -1,9 +1,9 @@
-import { useExperiment } from '../../context/experiment-context'
+import { useExperiment } from '../../context/experiment'
 import useStyles from './plots.style'
 import { TitleCard } from '../title-card/title-card'
 import { Tooltip, IconButton, Hidden } from '@mui/material'
 import ZoomOutMapIcon from '@mui/icons-material/ZoomOutMap'
-import { useGlobal } from '../../context/global-context'
+import { useGlobal } from '../../context/global'
 import { isUIBig } from '../../utility/ui-util'
 import { PlotList } from './plot-list'
 import { PlotItem } from './plot-item'
