+++ src/components/theme-selector/theme-selector.tsx
@@ -1,5 +1,5 @@
 import { Box, Button } from '@mui/material'
-import { useGlobal } from '../../context/global-context'
+import { useGlobal } from '../../context/global'
 import useStyles from './theme-selector.style'
 import { themes } from '../../theme/theme'
