+++ src/components/upload-csv-button.tsx
@@ -1,5 +1,5 @@
 import { IconButton, Input, Tooltip } from '@mui/material'
-import { useExperiment } from '../context/experiment-context'
+import { useExperiment } from '../context/experiment'
 import { DataPointType } from '../types/common'
 import { csvToDataPoints } from '../utility/converters'
 import PublishIcon from '@mui/icons-material/Publish'
