+++ src/context/experiment/experiment-context.tsx
@@ -1,17 +1,17 @@
 import * as React from 'react'
-import { versionInfo } from '../components/version-info'
-import { useLocalStorageReducer } from '../hooks/useLocalStorageReducer'
+import { versionInfo } from '../../components/version-info'
+import { useLocalStorageReducer } from '../../hooks/useLocalStorageReducer'
 import {
   Configuration,
   DefaultApi,
   OptimizerapiOptimizerRunRequest,
-} from '../../openapi'
-import { Dispatch, rootReducer } from '../reducers/reducers'
-import { initialState, State } from '../store'
-import { ExperimentResultType, ExperimentType } from '../types/common'
-import { calculateData, calculateSpace } from '../utility/converters'
-import { migrate } from '../utility/migration/migration'
-import { useGlobal } from './global-context'
+} from '../../../openapi'
+import { Dispatch, rootReducer } from './reducers'
+import { initialState, State } from '../../context/experiment/store'
+import { ExperimentResultType, ExperimentType } from '../../types/common'
+import { calculateData, calculateSpace } from '../../utility/converters'
+import { migrate } from '../../utility/migration/migration'
+import { useGlobal } from '../global/'
 
 const ExperimentContext = React.createContext<
   { state: State; dispatch: Dispatch; loading: boolean } | undefined
