+++ src/context/experiment/experiment-selectors.test.tsx
@@ -1,4 +1,4 @@
-import { initialState, State } from '../store'
+import { initialState, State } from '../../context/experiment/store'
 import { selectId, selectIsInitializing } from './experiment-selectors'
 
 describe('Experiment selectors', () => {
