+++ src/context/experiment/experiment-selectors.ts
@@ -1,4 +1,4 @@
-import { State } from '../store'
+import { State } from '../../context/experiment/store'
 
 export const selectExperiment = (state: State) => state.experiment
 export const selectId = (state: State) => selectExperiment(state).id
