+++ src/context/global/global-context.tsx
@@ -1,14 +1,9 @@
 import { Theme, ThemeProvider } from '@mui/material'
 import * as React from 'react'
-import ThemeSelector from '../components/theme-selector/theme-selector'
-import { useLocalStorageReducer } from '../hooks/useLocalStorageReducer'
-import {
-  State,
-  Dispatch,
-  initialState,
-  reducer,
-} from '../reducers/global-reducer'
-import { theme, themes, CustomTheme } from '../theme/theme'
+import ThemeSelector from '../../components/theme-selector/theme-selector'
+import { useLocalStorageReducer } from '../../hooks/useLocalStorageReducer'
+import { State, Dispatch, initialState, reducer } from './global-reducer'
+import { theme, themes, CustomTheme } from '../../theme/theme'
 
 declare module '@mui/styles/defaultTheme' {
   interface DefaultTheme extends Theme {}
