+++ src/hooks/useLocalStorageReducer.test.ts
@@ -1,7 +1,7 @@
 import { useLocalStorageReducer } from './useLocalStorageReducer'
 import { renderHook } from '@testing-library/react'
-import { initialState, State } from '../store'
-import { Action } from '../reducers/reducers'
+import { initialState, State } from '../context/experiment/store'
+import { Action } from '../context/experiment'
 
 describe('useLocalStorageReducer', () => {
   it('can be called once', () => {
