+++ src/utility/migration/migration.test.ts
@@ -5,7 +5,7 @@ import version4 from './data-formats/4.json'
 import version3 from './data-formats/3.json'
 import version2 from './data-formats/2.json'
 import version1 from './data-formats/1.json'
-import { emptyExperiment } from '../../store'
+import { emptyExperiment } from '../../context/experiment/store'
 import { ExperimentType } from '../../types/common'
 import fs from 'fs'
